requirements.txt
streamlit
python-pptx
groq